new file:   app/env.example.py

+++ app/routers/user.py
@@ -1,9 +1,11 @@
+"""Файл содержит все эндпоинты, который применяются к пользователям."""
+
 from fastapi import Depends, APIRouter, HTTPException
 from sqlalchemy.orm import Session
 from app.main import get_db
 from app.models.user import User
-
-
+from app.env import get_password_hash
+from app.auth.jwt_handler import get_current_user, check_user_permissions
 from app.schemas.user import UserBase, UserCreate, UserRead, UserUpdate
 
 
@@ -12,10 +14,9 @@ router = APIRouter()
 
 @router.post("/users", response_model=UserBase)
 async def create_user(user: UserCreate, db: Session = Depends(get_db)):
-    """
-    Создание пользователя
-    """
-    db_user = User(name=user.name, password=user.password, email=user.email)
+    """Создание пользователя."""
+    hashed_password = get_password_hash(user.password)
+    db_user = User(name=user.name, password=hashed_password, email=user.email)
     db.add(db_user)
     db.commit()
     db.refresh(db_user)
@@ -24,19 +25,20 @@ async def create_user(user: UserCreate, db: Session = Depends(get_db)):
 
 @router.get("/users", response_model=list[UserRead])
 def get_users(skip: int = 0, limit: int = 10, db: Session = Depends(get_db)):
-    """
-    Получить список пользователей
-    """
+    """Получить список пользователей."""
     users = db.query(User).offset(skip).limit(limit).all()
     return users
 
 
 @router.get("/users/{user_id}", response_model=UserRead)
-async def get_user(user_id: int, db: Session = Depends(get_db)):
-    """
-    Получить конкретного пользователя
-    """
-    user = db.query(User).filter(User.id == user_id).first
+async def get_user(
+    user_id: int,
+    current_user: UserRead = Depends(get_current_user),
+    db: Session = Depends(get_db),
+):
+    """Получить конкретного пользователя."""
+    check_user_permissions(user_id, current_user, db)
+    user = db.query(User).filter(User.id == user_id).first()
     if user:
         return user
     else:
@@ -45,11 +47,13 @@ async def get_user(user_id: int, db: Session = Depends(get_db)):
 
 @router.put("/users/{user_id}", response_model=UserRead)
 async def update_user(
-    user_id: int, user_update: UserUpdate, db: Session = Depends(get_db)
+    user_id: int,
+    user_update: UserUpdate,
+    current_user: UserRead = Depends(get_current_user),
+    db: Session = Depends(get_db),
 ):
-    """
-    Обновить данные о пользователе
-    """
+    """Обновить данные о пользователе."""
+    check_user_permissions(user_id, current_user, db)
     user = db.query(User).filter(User.id == user_id).first()
     if user is None:
         raise HTTPException(status_code=404, detail="Пользователь не найден")
@@ -62,12 +66,17 @@ async def update_user(
 
 
 @router.delete("/users/{user_id}")
-async def delete_user(user_id: int, db: Session = Depends(get_db)):
-    """
-    Удалить пользователя
-    """
+async def delete_user(
+    user_id: int,
+    current_user: UserRead = Depends(get_current_user),
+    db: Session = Depends(get_db),
+):
+    """Удалить пользователя."""
+    check_user_permissions(user_id, current_user, db)
     user = db.query(User).filter(User.id == user_id).first()
     if user is None:
         raise HTTPException(status_code=404, detail="Пользователь не найден")
     else:
         db.delete(user)
+        db.commit()
+        return {"message": "Пользователь удален"}
